sign in/sign out

+++ src/main/java/mhl/gochariot/model/Role.java
@@ -15,7 +15,7 @@ import lombok.NonNull;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long roleId;
+    private Integer roleId;
 
     @NonNull
     @Column(unique = true, nullable = false)
